jira-commention: Ignore words after "@" that are not email addresses
I noticed after writing the domain part of an email address in a Jira
comment that things following an "@" symbol that are not email addresses
shouldn't be filtered through the program. Update the regex matcher to
look for an "@" symbol in the text after the first "@" symbol to check
if it's an email address.